+++ src/boundary_conditions.jl
@@ -1,6 +1,4 @@
-#
-# Boundaries and Boundary Conditions
-#
+import GPUifyLoops: @launch, @loop, @unroll, @synchronize
 
 const coordinates = (:x, :y, :z)
 const dims = length(coordinates)
@@ -36,9 +34,9 @@ Adapt.adapt_structure(to, b::BoundaryCondition{C, A}) where {C<:BCType, A<:Abstr
     BoundaryCondition(C, Adapt.adapt(to, parent(b.condition)))
 
 """
-    CoordinateBoundaryConditions(c)
+    CoordinateBoundaryConditions
 
-Construct `CoordinateBoundaryCondition` to be applied along coordinate `c`, where
+Construct `CoordinateBoundaryConditions` to be applied along coordinate `c`, where
 `c` is `:x`, `:y`, or `:z`. A CoordinateBoundaryCondition has two fields
 `left` and `right` that store boundary conditions on the 'left' (negative side)
 and 'right' (positive side) of a given coordinate.
@@ -69,7 +67,7 @@ getbc(cbc::CBC, ::Val{:bottom}) = getfield(cbc, :right)
 getbc(cbc::CBC, ::Val{:top}) = getfield(cbc, :left)
 
 """
-    FieldBoundaryConditions
+    FieldBoundaryConditions <: FieldVector{dims, CoordinateBoundaryConditions}
 
 Construct `FieldBoundaryConditions` for a field.
 A FieldBoundaryCondition has `CoordinateBoundaryConditions` in
@@ -82,7 +80,7 @@ struct FieldBoundaryConditions <: FieldVector{dims, CoordinateBoundaryConditions
 end
 
 """
-    ModelBoundaryConditions
+    ModelBoundaryConditions <: FieldVector{nsolution, FieldBoundaryConditions}
 
 Construct a boundary condition type full of default
 `FieldBoundaryConditions` for u, v, w, T, S.
@@ -118,20 +116,17 @@ function ModelBoundaryConditions()
     return ModelBoundaryConditions(bcs...)
 end
 
-#
-# User API
-#
-
 #=
 Notes:
 
 - We assume that the boundary tendency has been previously calculated for
   a 'no-flux' boundary condition.
 
-  This means that boudnary conditions take the form of
-  an addition/subtraction to the tendency associated with a flux at point (A, A, I) below the bottom cell.
-  This paradigm holds as long as consider boundary conditions on (A, A, C) variables only, where A is
-  "any" of C or I.
+  This means that boudnary conditions take the form of an addition/subtraction
+  to the tendency associated with a flux at point `aaf` below the bottom cell.
+  This paradigm holds as long as consider boundary conditions on `aaf`
+  variables only, where a is "any" of c or f. See the src/operators/README for
+  more information on the naming convention for different grid point locations.
 
  - We use the physics-based convention that
 
@@ -140,7 +135,6 @@ Notes:
     and that
 
         tendency = ∂ϕ/∂t = Gϕ = - ∇ ⋅ flux
-
 =#
 
 const BC = BoundaryCondition
@@ -179,8 +173,7 @@ If `top_bc.condition` is a function, the function must have the signature
     Gϕ[i, j, 1] += κ * getbc(top_gradient, i, j, grid, t, iteration, u, v, w, T, S) / grid.Δz
 
 @inline apply_z_top_bc!(top_value::BC{<:Value}, i, j, grid, ϕ, Gϕ, κ, t, iteration, u, v, w, T, S) =
-    Gϕ[i, j, 1] += 2κ / grid.Δz * (
-        getbc(top_value, i, j, grid, t, iteration, u, v, w, T, S) - ϕ[i, j, 1])
+    Gϕ[i, j, 1] += 2κ / grid.Δz * (getbc(top_value, i, j, grid, t, iteration, u, v, w, T, S) - ϕ[i, j, 1])
 
 """
     apply_z_bottom_bc!(bottom_bc, i, j, grid, ϕ, Gϕ, κ, t, iteration, u, v, w, T, S)
@@ -199,5 +192,49 @@ If `bottom_bc.condition` is a function, the function must have the signature
     Gϕ[i, j, grid.Nz] -= κ * getbc(bottom_gradient, i, j, grid, t, iteration, u, v, w, T, S) / grid.Δz
 
 @inline apply_z_bottom_bc!(bottom_value::BC{<:Value}, i, j, grid, ϕ, Gϕ, κ, t, iteration, u, v, w, T, S) =
-    Gϕ[i, j, grid.Nz] -= 2κ / grid.Δz * (
-        ϕ[i, j, grid.Nz] - getbc(bottom_value, i, j, grid, t, iteration, u, v, w, T, S))
+    Gϕ[i, j, grid.Nz] -= 2κ / grid.Δz * (ϕ[i, j, grid.Nz] - getbc(bottom_value, i, j, grid, t, iteration, u, v, w, T, S))
+
+# Do nothing if both left and right boundary conditions are periodic.
+apply_bcs!(::CPU, ::Val{:x}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+apply_bcs!(::CPU, ::Val{:y}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+apply_bcs!(::CPU, ::Val{:z}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+
+apply_bcs!(::GPU, ::Val{:x}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+apply_bcs!(::GPU, ::Val{:y}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+apply_bcs!(::GPU, ::Val{:z}, Bx, By, Bz,
+    left_bc::BC{<:Periodic, T}, right_bc::BC{<:Periodic, T}, args...) where {T} = nothing
+
+# First, dispatch on coordinate.
+apply_bcs!(arch, ::Val{:x}, Bx, By, Bz, args...) =
+    @launch device(arch) threads=(Tx, Ty) blocks=(By, Bz) apply_x_bcs!(args...)
+apply_bcs!(arch, ::Val{:y}, Bx, By, Bz, args...) =
+    @launch device(arch) threads=(Tx, Ty) blocks=(Bx, Bz) apply_y_bcs!(args...)
+apply_bcs!(arch, ::Val{:z}, Bx, By, Bz, args...) =
+    @launch device(arch) threads=(Tx, Ty) blocks=(Bx, By) apply_z_bcs!(args...)
+
+"""
+    apply_z_bcs!(top_bc, bottom_bc, grid, ϕ, Gϕ, κ, closure, eos, g, t, iteration, u, v, w, T, S)
+
+Apply a top and/or bottom boundary condition to variable ϕ. Note that this kernel
+must be launched on the GPU with blocks=(Bx, By). If launched with blocks=(Bx, By, Bz),
+the boundary condition will be applied Bz times!
+"""
+function apply_z_bcs!(top_bc, bottom_bc, grid, ϕ, Gϕ, κ, closure, eos, g, t, iteration, u, v, w, T, S)
+    @loop for j in (1:grid.Ny; (blockIdx().y - 1) * blockDim().y + threadIdx().y)
+        @loop for i in (1:grid.Nx; (blockIdx().x - 1) * blockDim().x + threadIdx().x)
+
+               κ_top = κ(i, j, 1,       grid, closure, eos, g, u, v, w, T, S)
+            κ_bottom = κ(i, j, grid.Nz, grid, closure, eos, g, u, v, w, T, S)
+
+               apply_z_top_bc!(top_bc,    i, j, grid, ϕ, Gϕ, κ_top,    t, iteration, u, v, w, T, S)
+            apply_z_bottom_bc!(bottom_bc, i, j, grid, ϕ, Gϕ, κ_bottom, t, iteration, u, v, w, T, S)
+
+        end
+    end
+    @synchronize
+end
